refactor cookie module
