+++ src/auth/session/cookie/validation.ts
@@ -1,42 +1,31 @@
-import { getCookies, setCookie } from "@/encode/cookie";
-import { encodeJSON } from "@/encode/json";
 import type { BaseObject } from "@/utils/object";
 import { hasProperty, isObject } from "@/utils/object";
-import { isSome } from "@/utils/option";
 import type { Nullish } from "@/utils/type";
 import type { Session } from "next-auth";
 
-export function setSessionDataCookie(data: Session | null) {
-  return setCookie({
-    name: "session-data",
-    value: data ? encodeJSON(data) : "",
-    expiry: data ? new Date(data.expires) : new Date(),
-    path: "/",
-    sameSite: "lax",
-    secure: true,
-  });
-}
-
-const cookies = getCookies();
-export const cookieSession = isSome(cookies) ? assertSession(cookies.value["session-data"]) : null;
-
-function assertSession(value: unknown): Session | null {
+export function parseSession(value: unknown): Session | null {
   if (!isObject(value)) return null;
-  if (!hasUser(value)) return null;
+  if (!assertUser(value)) return null;
   if (!hasStringProperty(value, "expires")) return null;
   return value;
 }
 
-function hasUser(value: BaseObject): value is { user: Session["user"] } {
+export function parseWithExpires(value: unknown): Pick<Session, "expires"> | null {
+  if (!isObject(value)) return null;
+  if (!hasStringProperty(value, "expires")) return null;
+  return value;
+}
+
+function assertUser(value: BaseObject): value is { user: Session["user"] } {
   if (!hasProperty(value, "user")) return false;
   if (!isObject(value.user)) return false;
 
-  const user = assertUser(value.user);
+  const user = parseUser(value.user);
   if (!user) return false;
   return true;
 }
 
-function assertUser(value: BaseObject): Session["user"] | null {
+function parseUser(value: BaseObject): Session["user"] | null {
   if (!hasStringProperty(value, "id")) return null;
   if (!hasStringProperty(value, "name")) return null;
   if (!hasOptionalStringProperty(value, "email")) return null;
